+++ test/scrapers/imdbtv/testImdbTvShowSearch.cpp
@@ -16,7 +16,7 @@ TEST_CASE("ImdbTv returns valid search results", "[tv][ImdbTv][search]")
         auto* searchJob = new ImdbTvShowSearchJob(getImdbApi(), config);
         const auto scraperResults = test::searchTvScraperSync(searchJob).first;
 
-        REQUIRE(scraperResults.length() >= 8);
+        REQUIRE(scraperResults.length() >= 6);
         CHECK(scraperResults[0].title == "The Simpsons");
         CHECK(scraperResults[0].identifier.str() == "tt0096697");
         CHECK(scraperResults[0].released == QDate(1989, 1, 1)); // only year is set